pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "cli toll for get diff config file"
authors = ["Evgeny <fill1986@gmail.com>"]
packages = [
  { include = "gendiff_package" },
]


[tool.poetry.dependencies]
python = "^3.8"
prompt = "^0.4.1"
pytest = "^7.0.1"
pip = "^22.0.4"
install = "^1.3.5"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
pytest = "^7.0.1"
coverage = "^6.3.2"
pytest-cov = "^3.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
brain-calc = "brain_games.scripts.brain_calc:main"
gendiff = "gendiff_package.scripts.gendiff:main"
engine = "gendiff_package.scripts.engine:main"

